mini-pc init & log add timeinfo

+++ server/Common/Thread.h
@@ -17,7 +17,7 @@ public:
 	void RemoveTask(RoutineID nID);
 private:
 	//std::chrono::time_point<std::chrono::system_clock> m_threadTimer = std::chrono::system_clock::now();
-	int64_t m_nCurrencyTime = GetCurrencyTime();
+	long long m_nCurrencyTime = GetCurrencyTime();
 	std::mutex m_lock;
 	std::thread t;
 	std::vector<std::shared_ptr<Routine>> m_TickTaskList;
